requirements.txt
flask
openai==0.28
python-docx
PyPDF2